+++ module/applications/actor/base-sheet.mjs
@@ -1318,8 +1318,8 @@ export default class ActorSheet5e extends ActorSheetMixin(ActorSheet) {
       header = list?.previousElementSibling;
     }
     const mode = header?.dataset ?? {};
-    const isForcePower = Object.keys(CONFIG.SW5E.powerSchoolsForce).includes(itemData.system.school);
-    const isTechPower = Object.keys(CONFIG.SW5E.powerSchoolsTech).includes(itemData.system.school);
+    const isForcePower = itemData.system.school in CONFIG.SW5E.powerSchoolsForce;
+    const isTechPower = itemData.system.school in CONFIG.SW5E.powerSchoolsTech;
 
     // Determine the actor's power slot progressions, if any.
     const progs = Object.values(this.document.classes).reduce((acc, cls) => {